+++ src/types/related.ts
@@ -1,4 +1,4 @@
-import {FormatParam, PaginationParams, SortingParams, ViewParams} from "./base";
+import {FormatParam, PaginationParams, ViewParams} from "./base";
 
 export interface RelatedParams extends PaginationParams, ViewParams, FormatParam {
     media_id: string;